@@ -6,12 +6,12 @@ import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.ColorRes;
-import android.support.annotation.RequiresApi;
-import android.support.annotation.StringRes;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.ColorRes;
+import androidx.annotation.RequiresApi;
+import androidx.annotation.StringRes;
+import androidx.core.content.ContextCompat;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.view.animation.AccelerateInterpolator;
 import android.view.animation.AlphaAnimation;
@@ -21,7 +21,6 @@ import android.widget.Button;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
-import android.widget.TextView;
 
 import com.codemybrainsout.onboarder.utils.ShadowTransformer;
 import com.codemybrainsout.onboarder.views.CircleIndicatorView;
@@ -55,16 +54,16 @@ public abstract class AhoyOnboarderActivity extends AppCompatActivity implements
         setStatusBackgroundColor();
         hideActionBar();
 
-        parentLayout = (RelativeLayout) findViewById(R.id.parent_layout);
-        circleIndicatorView = (CircleIndicatorView) findViewById(R.id.circle_indicator_view);
-        btnSkip = (Button) findViewById(R.id.btn_skip);
-        buttonsLayout = (FrameLayout) findViewById(R.id.buttons_layout);
-        navigationControls = (FrameLayout) findViewById(R.id.navigation_layout);
-        ivNext = (ImageView) findViewById(R.id.ivNext);
-        ivPrev = (ImageView) findViewById(R.id.ivPrev);
-        backgroundImage = (ImageView) findViewById(R.id.background_image);
-        backgroundImageOverlay = (View) findViewById(R.id.background_image_overlay);
-        vpOnboarderPager = (ViewPager) findViewById(R.id.vp_pager);
+        parentLayout = findViewById(R.id.parent_layout);
+        circleIndicatorView = findViewById(R.id.circle_indicator_view);
+        btnSkip = findViewById(R.id.btn_skip);
+        buttonsLayout = findViewById(R.id.buttons_layout);
+        navigationControls = findViewById(R.id.navigation_layout);
+        ivNext = findViewById(R.id.ivNext);
+        ivPrev = findViewById(R.id.ivPrev);
+        backgroundImage = findViewById(R.id.background_image);
+        backgroundImageOverlay = findViewById(R.id.background_image_overlay);
+        vpOnboarderPager = findViewById(R.id.vp_pager);
         vpOnboarderPager.addOnPageChangeListener(this);
         btnSkip.setOnClickListener(this);
         ivPrev.setOnClickListener(this);
