@@ -12,6 +12,7 @@ public class FlowingGradient extends View {
 
     int duration;
     int draw;
+
     public FlowingGradient(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
         init();
@@ -19,10 +20,10 @@ public class FlowingGradient extends View {
 
     public FlowingGradient(Context context, AttributeSet attrs) {
         super(context, attrs);
-        TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.gradient,0, 0);
+        TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.gradient, 0, 0);
 
         draw = a.getResourceId(R.styleable.gradient_transition_drawable, R.drawable.translate);
-        duration = a.getInt(R.styleable.gradient_transition_duration,75);
+        duration = a.getInt(R.styleable.gradient_transition_duration, 75);
         init();
     }
 
@@ -36,8 +37,8 @@ public class FlowingGradient extends View {
         final AnimationDrawable frameAnimation = (AnimationDrawable) getBackground();
         frameAnimation.setEnterFadeDuration(duration);
         frameAnimation.setExitFadeDuration(duration);
-        post(new Runnable(){
-            public void run(){
+        post(new Runnable() {
+            public void run() {
                 frameAnimation.start();
             }
         });
